Move some files from setup package to better fitting ones
