+++ mobile/src/main/java/com/boswelja/devicemanager/watchmanager/ui/register/RegisterWatchViewModel.kt
@@ -5,7 +5,7 @@
  * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
  * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
  */
-package com.boswelja.devicemanager.setup.ui
+package com.boswelja.devicemanager.watchmanager.ui.register
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
@@ -18,7 +18,7 @@ import com.boswelja.devicemanager.watchmanager.item.Watch
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
-class WatchSetupViewModel @JvmOverloads constructor(
+class RegisterWatchViewModel @JvmOverloads constructor(
     application: Application,
     private val watchManager: WatchManager = WatchManager.get(application)
 ) : AndroidViewModel(application) {
