+++ mobile/src/main/res/navigation/setup_graph.xml
@@ -24,7 +24,7 @@
 
     <fragment
         android:id="@+id/watchSetupFragment"
-        android:name="com.boswelja.devicemanager.setup.ui.WatchSetupFragment"
+        android:name="com.boswelja.devicemanager.watchmanager.ui.register.RegisterWatchFragment"
         android:label="@string/register_watch_toolbar_title"
         tools:layout="@layout/fragment_watch_setup" />
